revert: #65 - refactor globals into singleton hooks
